Try with --watch

+++ src/index.ts
@@ -4,7 +4,11 @@ import * as typescript from 'typescript';
 import * as webpack from 'webpack';
 
 import * as constants from './constants';
-import { getEmitOutput, getTypeScriptInstance } from './instances';
+import {
+  getEmitOutput,
+  getTypeScriptInstance,
+  isReferencedFile
+} from './instances';
 import {
   LoaderOptions,
   LoaderOptionsCache,
@@ -271,19 +275,33 @@ function updateFileInCache(
   if (file === undefined) {
     file = instance.otherFiles.get(filePath);
     if (file !== undefined) {
-      instance.otherFiles.delete(filePath);
-      instance.files.set(filePath, file);
+      if (!isReferencedFile(instance, filePath)) {
+        instance.otherFiles.delete(filePath);
+        instance.files.set(filePath, file);
+        instance.changedFilesList = true;
+      }
     } else {
-      if (instance.watchHost !== undefined) {
+      if (
+        instance.watchHost !== undefined ||
+        instance.solutionBuilderHost !== undefined
+      ) {
         fileWatcherEventKind = instance.compiler.FileWatcherEventKind.Created;
       }
       file = { version: 0 };
-      instance.files.set(filePath, file);
+      if (!isReferencedFile(instance, filePath)) {
+        instance.files.set(filePath, file);
+        instance.changedFilesList = true;
+      } else {
+        instance.otherFiles.set(filePath, file);
+      }
     }
-    instance.changedFilesList = true;
   }
 
-  if (instance.watchHost !== undefined && contents === undefined) {
+  if (
+    (instance.watchHost !== undefined ||
+      instance.solutionBuilderHost !== undefined) &&
+    contents === undefined
+  ) {
     fileWatcherEventKind = instance.compiler.FileWatcherEventKind.Deleted;
   }
 
@@ -294,6 +312,7 @@ function updateFileInCache(
   //
   // See https://github.com/TypeStrong/ts-loader/issues/943
   if (
+    !isReferencedFile(instance, filePath) &&
     !instance.rootFileNames.has(filePath) &&
     // however, be careful not to add files from node_modules unless
     // it is allowed by the options.
@@ -308,7 +327,8 @@ function updateFileInCache(
     file.text = contents;
     instance.version!++;
     if (
-      instance.watchHost !== undefined &&
+      (instance.watchHost !== undefined ||
+        instance.solutionBuilderHost !== undefined) &&
       fileWatcherEventKind === undefined
     ) {
       fileWatcherEventKind = instance.compiler.FileWatcherEventKind.Changed;
@@ -319,6 +339,20 @@ function updateFileInCache(
     instance.hasUnaccountedModifiedFiles = true;
     instance.watchHost.invokeFileWatcher(filePath, fileWatcherEventKind);
     instance.watchHost.invokeDirectoryWatcher(path.dirname(filePath), filePath);
+  }
+
+  if (
+    instance.solutionBuilderHost !== undefined &&
+    fileWatcherEventKind !== undefined
+  ) {
+    instance.solutionBuilderHost.invokeFileWatcher(
+      filePath,
+      fileWatcherEventKind
+    );
+    instance.solutionBuilderHost.invokeDirectoryWatcher(
+      path.dirname(filePath),
+      filePath
+    );
   }
 
   // push this file to modified files hash.
@@ -338,36 +372,38 @@ function getEmit(
 ) {
   const outputFiles = getEmitOutput(instance, filePath, loaderContext);
 
-  loaderContext.clearDependencies();
-  loaderContext.addDependency(rawFilePath);
+  if (!isReferencedFile(instance, filePath)) {
+    loaderContext.clearDependencies();
+    loaderContext.addDependency(rawFilePath);
 
-  const allDefinitionFiles = [...instance.files.keys()].filter(defFilePath =>
-    defFilePath.match(constants.dtsDtsxOrDtsDtsxMapRegex)
-  );
-
-  // Make this file dependent on *all* definition files in the program
-  const addDependency = loaderContext.addDependency.bind(loaderContext);
-  allDefinitionFiles.forEach(addDependency);
-
-  // Additionally make this file dependent on all imported files
-  const fileDependencies = instance.dependencyGraph[filePath];
-  const additionalDependencies =
-    fileDependencies === undefined
-      ? []
-      : fileDependencies.map(({ originalFileName }) => originalFileName);
-
-  if (additionalDependencies.length > 0) {
-    additionalDependencies.forEach(addDependency);
-  }
-
-  loaderContext._module.buildMeta.tsLoaderDefinitionFileVersions = allDefinitionFiles
-    .concat(additionalDependencies)
-    .map(
-      defFilePath =>
-        defFilePath +
-        '@' +
-        (instance.files.get(defFilePath) || { version: '?' }).version
+    const allDefinitionFiles = [...instance.files.keys()].filter(defFilePath =>
+      defFilePath.match(constants.dtsDtsxOrDtsDtsxMapRegex)
     );
+
+    // Make this file dependent on *all* definition files in the program
+    const addDependency = loaderContext.addDependency.bind(loaderContext);
+    allDefinitionFiles.forEach(addDependency);
+
+    // Additionally make this file dependent on all imported files
+    const fileDependencies = instance.dependencyGraph[filePath];
+    const additionalDependencies =
+      fileDependencies === undefined
+        ? []
+        : fileDependencies.map(({ originalFileName }) => originalFileName);
+
+    if (additionalDependencies.length > 0) {
+      additionalDependencies.forEach(addDependency);
+    }
+
+    loaderContext._module.buildMeta.tsLoaderDefinitionFileVersions = allDefinitionFiles
+      .concat(additionalDependencies)
+      .map(
+        defFilePath =>
+          defFilePath +
+          '@' +
+          (instance.files.get(defFilePath) || { version: '?' }).version
+      );
+  }
 
   const outputFile = outputFiles
     .filter(file => file.name.match(constants.jsJsx))
